+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 This update introduces several improvements and bug fixes.
 
+### Added
+ - Add optional UUID fields for blacklist, evidence, and file attachments in AuditLog
+
 ### Changed
  - Changed audit_log table structure by adding optional fields for blacklist, evidence, and file attachments; update
    timestamp index in audit_log.sql